+++ aquadoggo/src/materializer/tasks/reduce.rs
@@ -21,7 +21,13 @@ pub async fn reduce_task(context: Context, input: TaskInput) -> TaskResult<TaskI
     debug!("Working on reduce task {:#?}", input);
 
     // Find out which document we are handling
-    let document_id = resolve_document_id(&context, &input).await?;
+    let document_id = match resolve_document_id(&context, &input).await? {
+        Some(document_id) => Ok(document_id),
+        None => {
+            debug!("No document found for this view, exit without dispatching any other tasks");
+            return Ok(None);
+        }
+    }?;
 
     // Get all operations for the requested document
     let operations = context
@@ -65,27 +71,27 @@ pub async fn reduce_task(context: Context, input: TaskInput) -> TaskResult<TaskI
 async fn resolve_document_id(
     context: &Context,
     input: &TaskInput,
-) -> Result<DocumentId, TaskError> {
+) -> Result<Option<DocumentId>, TaskError> {
     match (&input.document_id, &input.document_view_id) {
         // The `DocumentId` is already given, we don't have to do anything
-        (Some(document_id), None) => Ok(document_id.to_owned()),
+        (Some(document_id), None) => Ok(Some(document_id.to_owned())),
 
         // A `DocumentViewId` is given, let's find out its document id
         (None, Some(document_view_id)) => {
             // @TODO: We can skip this step if we implement:
             // https://github.com/p2panda/aquadoggo/issues/148
+            debug!("Find document for view with id: {}", document_view_id);
             let operation_id = document_view_id.clone().into_iter().next().unwrap();
-            match context
+            context
                 .store
                 .get_document_by_operation_id(&operation_id)
                 .await
-                .map_err(|_| TaskError::Critical)?
-            {
-                Some(document_id) => Ok(document_id),
-                None => Err(TaskError::Critical),
-            }
+                .map_err(|err| {
+                    debug!("Fatal error getting document_id from storage");
+                    debug!("{}", err);
+                    TaskError::Critical
+                })
         }
-
         // None or both have been provided which smells like a bug
         (_, _) => Err(TaskError::Critical),
     }
@@ -189,27 +195,19 @@ async fn reduce_document(
 
 #[cfg(test)]
 mod tests {
-    use std::convert::TryFrom;
-
     use p2panda_rs::document::{DocumentBuilder, DocumentId, DocumentViewId};
-    use p2panda_rs::entry::{sign_and_encode, Entry};
-    use p2panda_rs::identity::Author;
-    use p2panda_rs::operation::{
-        AsVerifiedOperation, OperationEncoded, OperationValue, VerifiedOperation,
-    };
-    use p2panda_rs::storage_provider::traits::{
-        AsStorageEntry, EntryStore, OperationStore, StorageProvider,
-    };
+    use p2panda_rs::operation::{AsVerifiedOperation, OperationValue};
+    use p2panda_rs::storage_provider::traits::OperationStore;
     use p2panda_rs::test_utils::constants::TEST_SCHEMA_ID;
-    use p2panda_rs::test_utils::fixtures::{operation, operation_fields};
+    use p2panda_rs::test_utils::fixtures::{
+        operation, operation_fields, random_document_id, random_document_view_id,
+    };
     use rstest::rstest;
 
     use crate::config::Configuration;
     use crate::context::Context;
-    use crate::db::stores::test_utils::{test_db, TestDatabase, TestDatabaseRunner};
-    use crate::db::stores::StorageEntry;
+    use crate::db::stores::test_utils::{send_to_store, test_db, TestDatabase, TestDatabaseRunner};
     use crate::db::traits::DocumentStore;
-    use crate::graphql::client::EntryArgsRequest;
     use crate::materializer::tasks::reduce_task;
     use crate::materializer::TaskInput;
 
@@ -218,7 +216,7 @@ mod tests {
         #[from(test_db)]
         #[with(
             2,
-            20,
+            5,
             false,
             TEST_SCHEMA_ID.parse().unwrap(),
             vec![("username", OperationValue::Text("panda".into()))],
@@ -254,7 +252,6 @@ mod tests {
         runner.with_db_teardown(|db: TestDatabase| async move {
             let document_id = db.documents.first().unwrap();
             let key_pair = db.key_pairs.first().unwrap();
-            let author = Author::try_from(key_pair.public_key().to_owned()).unwrap();
 
             let context = Context::new(db.store.clone(), Configuration::default());
             let input = TaskInput::new(Some(document_id.clone()), None);
@@ -264,48 +261,20 @@ mod tests {
             assert!(reduce_task(context.clone(), input.clone()).await.is_ok());
 
             // Now we create and insert an UPDATE operation for this document.
-            let entry_args = db
-                .store
-                .get_entry_args(&EntryArgsRequest {
-                    author,
-                    document: Some(document_id.clone()),
-                })
-                .await
-                .unwrap();
-
-            let operation = operation(
-                Some(operation_fields(vec![(
-                    "username",
-                    OperationValue::Text("meeeeeee".to_string()),
-                )])),
-                Some(document_id.as_str().parse().unwrap()),
-                None,
-            );
-
-            let entry = Entry::new(
-                &entry_args.log_id,
-                Some(&operation),
-                entry_args.skiplink.as_ref(),
-                entry_args.backlink.as_ref(),
-                &entry_args.seq_num,
+            let (_, _) = send_to_store(
+                &db.store,
+                &operation(
+                    Some(operation_fields(vec![(
+                        "username",
+                        OperationValue::Text("meeeeeee".to_string()),
+                    )])),
+                    Some(document_id.as_str().parse().unwrap()),
+                    None,
+                ),
+                Some(document_id),
+                key_pair,
             )
-            .unwrap();
-
-            let entry_signed = sign_and_encode(&entry, key_pair).unwrap();
-            let operation_encoded = OperationEncoded::try_from(&operation).unwrap();
-
-            db.store
-                .insert_entry(StorageEntry::new(&entry_signed, &operation_encoded).unwrap())
-                .await
-                .unwrap();
-
-            let verified_operation =
-                VerifiedOperation::new_from_entry(&entry_signed, &operation_encoded).unwrap();
-
-            db.store
-                .insert_operation(&verified_operation, document_id)
-                .await
-                .unwrap();
+            .await;
 
             // This should now find the new UPDATE operation and perform an update on the document
             // in the documents table.
@@ -470,6 +439,28 @@ mod tests {
             let input = TaskInput::new(document_id, document_view_id);
 
             assert!(reduce_task(context.clone(), input).await.is_err());
+        });
+    }
+
+    #[rstest]
+    fn does_not_error_when_document_missing(
+        #[from(test_db)]
+        #[with(0, 0)]
+        runner: TestDatabaseRunner,
+        #[from(random_document_id)] document_id: DocumentId,
+        #[from(random_document_view_id)] document_view_id: DocumentViewId,
+    ) {
+        // Prepare empty database.
+        runner.with_db_teardown(|db: TestDatabase| async move {
+            let context = Context::new(db.store.clone(), Configuration::default());
+
+            // Dispatch a reduce task for a document which doesn't exist by it's document id.
+            let input = TaskInput::new(Some(document_id), None);
+            assert!(reduce_task(context.clone(), input).await.is_ok());
+
+            // Dispatch a reduce task for a document which doesn't exist by it's document view id.
+            let input = TaskInput::new(None, Some(document_view_id));
+            assert!(reduce_task(context.clone(), input).await.is_ok());
         });
     }
 }